Install improvements

+++ manage.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 import os
-from flask_script import Manager, Shell, Server, prompt_pass
+import datetime
+from dateutil.relativedelta import relativedelta
+from random import randint
+from flask_script import Manager, Shell, Server, prompt_pass, prompt_bool
 from flask_migrate import MigrateCommand, upgrade
 from comport.content.defaults import ChartBlockDefaults
 
@@ -29,7 +32,8 @@ if os.environ.get("COMPORT_ENV") == 'prod':
     config_object = ProdConfig
 else:
     # No Slack webhook URL for testing
-    del(os.environ['SLACK_WEBHOOK_URL'])
+    if 'SLACK_WEBHOOK_URL' in os.environ:
+        del(os.environ['SLACK_WEBHOOK_URL'])
     config_object = DevConfig
 
 # create the app
@@ -215,22 +219,72 @@ def add_new_blocks():
 
 @manager.command
 def test_client():
+    ''' Erase the database and load in a full suite of test data
+    '''
+    if not prompt_bool("Are you sure you want to destroy and recreate Comport's database?"):
+        return
+
     delete_everything()
-    department = Department.create(name="Busy Town Police Department", short_name="BTPD", load_defaults=True)
-    user = User.create(username="user", email="email2@example.com", password="password", active=True, is_admin=True)
+
+    # create a fake PD and admin user
+    department = Department.create(name="Izquierda Metropolitan Police Department", short_name="IMPD", load_defaults=True)
+    user = User.create(username="user", email="user@example.com", password="password", active=True, is_admin=True)
     user.departments.append(department)
     user.save()
 
-    test_client = JSONTestClient()
-    # missing_data_mutator = MissingDataMutator()
-    # fuzzed_data_mutator = FuzzedDataMutator()
-    # known_bad_data_mutator = KnownBadDataMutator()
-    # empty_data_mutator = EmptyDataMutator()
-    # casing_mutator = CasingMutator()
-    # condenisng_date_mutator = CondenisngDateMutator()
-    # gap_date_mutator = GapDateMutator()
+    # create some fake officer out on service data
+    date_now = datetime.datetime.now()
+    date_step = date_now - relativedelta(months=30)
+    while date_step.year < date_now.year or date_step.month < date_now.month:
+        DenominatorValue.create(
+            department_id=department.id,
+            month=date_step.month,
+            year=date_step.year,
+            officers_out_on_service=(100000 + (randint(0, 46000) - 23000))
+        )
+        date_step = date_step + relativedelta(months=1)
 
-    test_client.run(department, [])
+    # create some fake demographic data
+    demo_template = [
+        dict(race="Asian", city_factor=0.0194, dept_factor=0.0013),
+        dict(race="Black", city_factor=0.2452, dept_factor=0.1402),
+        dict(race="Hispanic", city_factor=0.0861, dept_factor=0.0253),
+        dict(race="Other", city_factor=0.0699, dept_factor=0.0101),
+        dict(race="White", city_factor=0.5794, dept_factor=0.8231)
+    ]
+
+    # for the city
+    city_population = 100000 + round(100000 * ((randint(0, 16) / 100) - .08))
+    for value in demo_template:
+        DemographicValue.create(
+            department_id=department.id,
+            race=value["race"],
+            count=round(city_population * value["city_factor"]),
+            department_value=False
+        )
+
+    # for the department
+    dept_population = 1500 + round(1500 * ((randint(0, 16) / 100) - .08))
+    for value in demo_template:
+        DemographicValue.create(
+            department_id=department.id,
+            race=value["race"],
+            count=round(dept_population * value["dept_factor"]),
+            department_value=True
+        )
+
+    # create a JSON test client and run it
+    test_client = JSONTestClient()
+    mutations = []
+    # mutations.append(MissingDataMutator())
+    # mutations.append(FuzzedDataMutator())
+    # mutations.append(KnownBadDataMutator())
+    # mutations.append(EmptyDataMutator())
+    # mutations.append(CasingMutator())
+    # mutations.append(CondenisngDateMutator())
+    # mutations.append(GapDateMutator())
+
+    test_client.run(department, mutations)
 
 
 manager.add_command('server', Server())
